backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
httpx==0.27.0
python-dotenv>=1.0.0
asyncpg>=0.29.0
PyJWT>=2.8.0
bcrypt>=4.0.0
boto3>=1.34.0
oss2>=2.18.0
email-validator>=2.0.0
imageio-ffmpeg>=0.4.9
